change threshold_double to float

+++ include/LightGBM/dataset.h
@@ -509,7 +509,7 @@ class Dataset {
   }
 
   // given a real threshold, find the closest threshold bin
-  inline uint32_t BinThreshold(int i, double threshold_double) const {
+  inline uint32_t BinThreshold(int i, float threshold_double) const {
     const int group = feature2group_[i];
     const int sub_feature = feature2subfeature_[i];
     return feature_groups_[group]->bin_mappers_[sub_feature]->ValueToBin(threshold_double);
